Feature/23/add null check

+++ server/src/service/ProductService.ts
@@ -15,45 +15,65 @@ export class ProductService {
     }
 
     async getAll(): Promise<Result> {
-        let result = new Result();
-        result.data = await this.productRepository.getAll();
-        return result;
+        const data = await this.productRepository.getAll();
+        return new Result({ data: data });
     }
 
     async getById(id: number): Promise<Result> {
-        let result = new Result();
-        result.data = await this.productRepository.getById(id);
-        if (!result.data) {
-            result.message = "Product not found";
+        if (!id) {
+            return new Result({ message: "Missing fields" });
         }
-        return result;
+        const data = await this.productRepository.getById(id);
+        if (!data) {
+            return new Result({ message: "Product not found" });
+        }
+        return new Result({ data: data });
     }
 
     async add(product: Product): Promise<Result> {
-        let result = new Result();
-        result.data = await this.productRepository.save(product);
-        return result;
+        if (
+            !product.name ||
+            !product.category ||
+            !product.image ||
+            !product.price ||
+            !product.description ||
+            !product.category
+        ) {
+            return new Result({ message: "Missing fields" });
+        }
+        const data = await this.productRepository.save(product);
+        return new Result({ data: data });
     }
 
     async update(product: Product): Promise<Result> {
-        let result = new Result();
+        if (
+            !product.name ||
+            !product.category ||
+            !product.image ||
+            !product.price ||
+            !product.description ||
+            !product.category ||
+            !product.id
+        ) {
+            return new Result({ message: "Missing fields" });
+        }
         const productTemp = await this.productRepository.getById(product.id);
         if (!productTemp) {
-            result.message = "Product not found";
-        } else {
-            await this.productRepository.update(product.id, product);
+            return new Result({ message: "Product not found" });
         }
-        return result;
+        await this.productRepository.update(product.id, product);
+        return new Result({});
     }
 
     async delete(id: number): Promise<Result> {
-        let result = new Result();
+        if (!id) {
+            return new Result({ message: "Missing fields" });
+        }
         const product = await this.productRepository.getById(id);
         if (!product) {
-            result.message = "Product not found";
-        } else {
-            await this.productRepository.delete(id);
+            return new Result({ message: "Product not found" });
         }
-        return result;
+        await this.productRepository.delete(id);
+        return new Result({});
     }
 }
